% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generations.R
\name{add_years}
\alias{add_years}
\title{Add years to generation name}
\usage{
add_years(gens, years_sep, years_range_sep)
}
\arguments{
\item{gens}{a vector}

\item{years_sep}{a string}

\item{years_range_sep}{a string}
}
\description{
Takes a vector of generation names (short or long) and two integers. Adds \verb{ (x–y)}
to the end of the names. The space and en-dash can be modified with the two
arguments.
}
\examples{
add_years(c("Millennial", "Gen X"))
}
\keyword{internal}
