% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_df.R
\docType{data}
\name{gen_df}
\alias{gen_df}
\title{Generations Lookup Table}
\format{
An object of class \code{data.frame} with 25 rows and 3 columns.
}
\usage{
gen_df
}
\description{
Generations Lookup Table
}
\keyword{datasets}
