% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_info.R
\name{get_prev_gen}
\alias{get_prev_gen}
\alias{get_next_gen}
\title{Get the previous generation}
\usage{
get_prev_gen(gen)

get_next_gen(gen)
}
\arguments{
\item{gen}{a string, exactly matching a generation's name}
}
\value{
a string. If the generation is out of range (such as the generation
after Gen Z, it returns \code{NA}.)
}
\description{
Given a generation's name, find what the previous or next generation is.
}
\examples{
get_prev_gen("Millennial")
get_prev_gen("Aurthurian")
get_next_gen("Silent")
get_next_gen("Gen Z")
}
