% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{redefine_generation}
\alias{redefine_generation}
\title{Redefine when a generation starts and ends}
\usage{
redefine_generation(generation, start, end)
}
\arguments{
\item{generation}{a string, exactly matching a generation's name}

\item{start}{an integer, representing the year you want \code{generation} to
start at}

\item{end}{an integer, representing thet year you want \code{generation} to
end at.}
}
\description{
Not all people agree on when a generation started and ended. This is especially
true with the most recent few generations. The years I have chosen
come from Howe & Strauss' book. This function lets you redefine the start
and end years of a generation to fit your personal tastes.
}
\details{
The changes here modify the underlying data and will be reflected in any
subsequent reference to the generational data. You can view what the current
dataset is with \code{show_generations}. Any change you make can be reset using
\code{reset_generations}.
}
\examples{
redefine_generation("Millennial", 1980, 1997)
show_generations()
}
