% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{rename_generation}
\alias{rename_generation}
\alias{use_zoomer}
\alias{use_gen_z}
\alias{use_gen_y}
\alias{use_millennial}
\alias{use_13th}
\alias{use_gen_x}
\alias{use_baby_boom}
\alias{use_boomer}
\title{Rename a generation}
\usage{
rename_generation(old_name, new_name)

use_zoomer()

use_gen_z()

use_gen_y()

use_millennial()

use_13th()

use_gen_x()

use_baby_boom()

use_boomer()
}
\arguments{
\item{old_name}{a string, exactly matching a generation's name}

\item{new_name}{a string, representing what you want that generation to be
called}
}
\description{
Not all people refer to generations the same way. The names I have chosen
come from Howe & Strauss' book, though I have taken some liberty with the
most recent four generations. This family of functions lets you rename
a generation to fit your personal tastes.
}
\details{
The main function is \code{rename_generation}. All others are shortcuts for
some of the more common changes I expect people to make.

The changes here modify the underlying data and will be reflected in any
subsequent reference to the generational data. You can view what the current
dataset is with \code{show_generations}. Any change you make can be reset using
\code{reset_generations}.
}
\examples{
rename_generation("Gen Z", "Zoomer")
rename_generation("Boomer", "Baby Boom")
show_generations()

use_gen_z()
use_zoomer()
use_gen_y()
use_millennial()
use_13th()
use_gen_x()
use_boomer()
use_baby_boom()
}
