% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{reset_generations}
\alias{reset_generations}
\title{Reset underlying generational data}
\usage{
reset_generations()
}
\description{
Changes to the underlying dataset can be made with \code{rename_generation}
or \code{redefine_generation}. This function undoes all these changes and
reverts the data back to its out-of-the-box names and definitions.
}
\seealso{
\code{\link{rename_generation}}, \code{\link{redefine_generation}}
}
