% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{show_generations}
\alias{show_generations}
\title{Show generational data}
\usage{
show_generations()
}
\description{
Displays the generation names, start years, and end years.
}
\examples{
show_generations()
}
