% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_start}
\alias{get_start}
\alias{get_end}
\title{Get a generation's start or end year}
\usage{
get_start(generation)

get_end(generation)
}
\arguments{
\item{generation}{a string, exactly matching a generation's name}
}
\description{
Get a generation's start or end year
}
\examples{
get_start("Millennial")
get_start(c("Millennial", "Lost"))
get_end("Millennial")
get_end(c("Millennial", "Lost"))
}
