% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generations.R
\name{short_generation_names}
\alias{short_generation_names}
\title{Get shortened generation names}
\usage{
short_generation_names(gens)
}
\arguments{
\item{gens}{a vector}
}
\description{
For internal use only. Mostly used within \code{add_years()} so that
\code{generations()} words properly when \code{full_names} and \code{years}
are both specified.
}
